third_party/blink/web_tests/platform/generic/fast/forms/autocomplete-expected.txt

This tests that the 'autocomplete' attribute of form and input elements is bound to a JavaScript property.
PASS ('autocomplete' in form) is true
PASS ('autocomplete' in field) is true
PASS form.getAttribute('autocomplete') is null
PASS field.getAttribute('autocomplete') is null
PASS form.autocomplete is "on"
PASS field.autocomplete is ""
PASS form.autocomplete = 'off'; form.getAttribute('autocomplete') is "off"
PASS field.autocomplete = 'off'; field.getAttribute('autocomplete') is "off"
PASS form.setAttribute('autocomplete', 'on'); form.autocomplete is "on"
PASS field.setAttribute('autocomplete', 'on'); field.autocomplete is "on"
PASS form.setAttribute('autocomplete', 'indeterminate'); form.autocomplete is "on"
PASS field.setAttribute('autocomplete', 'indeterminate'); field.autocomplete is ""
PASS successfullyParsed is true

TEST COMPLETE
